Some very rough and ready testing to try to find out why you can upload data into the new scicat backend with pyscicat, but you cannot read it.
I wanted to check if this was an auth error or an error parsing the RawDataset model. I created two tests :
1) `test_upload_dataset_requests()` which uploaded a RawDataset model via the requests protocol
2) `test_upload_dataset_sci_cli()` which uploaded a dictionary using the ScicatClient. I had to change the _send_to_scicat method to allow upload of dictionaries.

I found that method 2 can ingest to scicat in a way that the API can read the data. Method 1 will ingest the dataset but the Scicat API cannot read it.